+++ tests/conftest.py
@@ -9,7 +9,7 @@ import pytest
 import requests_mock
 from faker import Faker
 
-from sekoia_automation.configuration import get_configuration
+from sekoia_automation.configuration.filesystem import FileSystemConfiguration
 from sekoia_automation import constants
 from sekoia_automation import storage as storage_module
 from sekoia_automation.module import Module
@@ -53,19 +53,19 @@ def mock_volume():
 
 @pytest.fixture
 def config():
-    return get_configuration()
+    return FileSystemConfiguration()
 
 
 @pytest.fixture
 def config_storage(config):
-    old_config_storage = config.VOLUME_PATH
-    config.VOLUME_PATH = mkdtemp()
-    storage_module.VOLUME_PATH = config.VOLUME_PATH
+    old_config_storage = FileSystemConfiguration.VOLUME_PATH
+    FileSystemConfiguration.VOLUME_PATH = mkdtemp()
+    storage_module.VOLUME_PATH = FileSystemConfiguration.VOLUME_PATH
 
-    yield Path(config.VOLUME_PATH)
+    yield Path(FileSystemConfiguration.VOLUME_PATH)
 
-    rmtree(config.VOLUME_PATH)
-    config.VOLUME_PATH = old_config_storage
+    rmtree(FileSystemConfiguration.VOLUME_PATH)
+    FileSystemConfiguration.VOLUME_PATH = old_config_storage
 
 
 @pytest.fixture