+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Check an API key is provided in the configuration, and add it to the Authorization header,
+  when no authentication method is set in an action
+
 ## 1.20.3a1 - 2025-09-23
 
 ### Added